requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
